applies a fix for destroy provisioner failures in terraform >= 0.13.x
the fix is one of the workarounds proposed in this issue:
https://github.com/hashicorp/terraform/issues/23679 (use self). while
this workaround has risk mentioned in the issue comments, karrots views
k8s clusters as immutable and ephemeral so we don't need destroy.
karrots' philosophy is to never change a running cluster. any destroy
operation would be against the entire cluster in one go, so fluxcd
disappears when the cluster disappears (removing the need for fluxcd to
handle any destroy cases). so this "fix" is really just to get fluxcd to
work for 'apply' given tf versions >= 0.13.x